minor cleanup

+++ src/linkgene.cpp
@@ -1,17 +1,17 @@
 /*
- Copyright 2001 The University of Texas at Austin
+  Copyright 2001 The University of Texas at Austin
 
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
 
-       http://www.apache.org/licenses/LICENSE-2.0
+  http://www.apache.org/licenses/LICENSE-2.0
 
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
 */
 #include "linkgene.h"
 
@@ -20,11 +20,11 @@
 using namespace NEAT;
 
 LinkGene::LinkGene(double w,
-           int inode_id,
-           int onode_id,
-           bool recur,
-           double innov,
-           double mnum) {
+                   int inode_id,
+                   int onode_id,
+                   bool recur,
+                   int innov,
+                   double mnum) {
     _weight = w;
     _in_node_id = inode_id;
     _out_node_id = onode_id;
@@ -40,12 +40,12 @@ LinkGene::LinkGene(double w,
 
 //Construct a gene with a trait
 LinkGene::LinkGene(int trait_id,
-           double w,
-           int inode_id,
-           int onode_id,
-           bool recur,
-           double innov,
-           double mnum) {
+                   double w,
+                   int inode_id,
+                   int onode_id,
+                   bool recur,
+                   int innov,
+                   double mnum) {
     _weight = w;
     _in_node_id = inode_id;
     _out_node_id = onode_id;
@@ -59,9 +59,9 @@ LinkGene::LinkGene(int trait_id,
 }
 
 LinkGene::LinkGene(LinkGene *g,
-           int trait_id,
-           int inode_id,
-           int onode_id) {
+                   int trait_id,
+                   int inode_id,
+                   int onode_id) {
     _weight = g->_weight;
     _in_node_id = inode_id;
     _out_node_id = onode_id;
